ai-travel-agent-main/requirements.txt
# Core dependencies
langchain>=0.2.0
langchain-openai>=0.1.7
langchain-core>=0.2.0
python-dotenv>=1.0.0

# Email
sendgrid>=6.10.0

# Search & APIs
google-search-results>=2.4.2
requests>=2.31.0

# ML/Data Science
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Utilities
joblib>=1.3.0